+++ src/sql/__init__.py
@@ -1,6 +1,6 @@
 from .magic import RenderMagic, SqlMagic, load_ipython_extension
 
-__version__ = "0.5.7dev"
+__version__ = "0.6.0dev"
 
 
 __all__ = [