+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGELOG
 
-## 0.5.7dev
+## 0.6.0dev
 
 * [Feature] Adds `%%config SqlMagic.autopolars = True` ([#138](https://github.com/ploomber/jupysql/issues/138))
+* [API Change] Drops support for old versions of IPython (removed imports from `IPython.utils.traitlets`)
 
 ## 0.5.6 (2023-02-16)